WIP to make dtbuilder work on 64bit.
Fixes many, though not all 64bit-warnings.  In lots of places, pointers are
cast to ints to be then used as array subscripts.  The only way to deal with
this is to change them to long.  Additionally, use calloc() to allocate the
int_array in istr.c and drop the (wrong) macro patch to istr.h.  Should make
dtbuilder work on 32bit again.

+++ cde/programs/dtappbuilder/src/ab/abobj_layout.c
@@ -110,7 +110,7 @@ abobj_calculate_new_layout(
     Position    x, y;
     Dimension   width, height;
     Dimension   p_width, p_height;
-    int		gridpos;
+    long	gridpos;
 
     if ((attachments = obj->attachments) == NULL)
         return -1;
